pytorch/natural_language_processing/text_anomaly_detection/requirements.txt
click==7.0
joypy==0.2.1
matplotlib==3.1.1
nltk==3.4.4
numpy==1.17.0
pandas==0.25.0
pytorch-nlp==0.4.1
pytorch-pretrained-bert==0.6.1
regex==2018.1.10
scikit-learn==0.21.2
seaborn==0.9.0
spacy==2.0.18
torch==1.9.0
torchtext==0.10.0
tensorboardX==2.4